# specs/001-local-llm-webapp/contracts/api-spec.yaml
openapi: 3.1.0
info:
  title: Local LLM Web Application API
  description: |
    REST API for air-gapped Local LLM web application for local government employees.

    **Key Features**:
    - User authentication and session management
    - Conversational AI with streaming responses
    - Document upload and analysis
    - Conversation history management
    - Administrator dashboard and user management

    **Base URL**: `http://<server-ip>:8000/api/v1`
  version: 1.0.0
  contact:
    name: Development Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: http://{server-ip}:8000/api/v1
    description: Production server (air-gapped)
    variables:
      server-ip:
        default: '192.168.1.100'
        description: Internal network IP address
...
paths:
...
  /chat/send:
    post:
      tags: [Chat]
      summary: Send message and get LLM response
      description: Submit user query and receive complete LLM response (non-streaming)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversation_id, content]
              properties:
                conversation_id:
                  type: string
                  format: uuid
                  description: Existing conversation ID or null for new conversation
                  nullable: true
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
                  example: "행정 업무에서 공문서를 작성할 때 주의사항이 무엇인가요?"
                document_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Optional document IDs for context
                  nullable: true
      responses:
        '200':
          description: LLM response generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                    format: uuid
                  message:
                    $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /chat/stream:
    post:
      tags: [Chat]
      summary: Send message with streaming response
      description: Submit user query and receive LLM response as Server-Sent Events (SSE)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversation_id, content]
              properties:
                conversation_id:
                  type: string
                  format: uuid
                  nullable: true
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
                document_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  nullable: true
      responses:
        '200':
          description: Streaming response (Server-Sent Events)
          content:
            text/event-stream:
              schema:
                type: object
                properties:
                  event:
                    type: string
                    enum: [token, done, error]
                  data:
                    oneOf:
                      - type: object
                        description: Token event
                        properties:
                          token:
                            type: string
                      - type: object
                        description: Done event
                        properties:
                          conversation_id:
                            type: string
                            format: uuid
                          message_id:
                            type: string
                            format: uuid
                      - type: object
                        description: Error event
                        properties:
                          error:
                            type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /documents:
    get:
      tags: [Documents]
      summary: List user's documents
      description: Retrieve user's uploaded documents
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Documents]
      summary: Upload document
      description: Upload PDF, TXT, or DOCX file for analysis
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file (max 50MB)
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File too large (max 50MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /documents/{document_id}:
    get:
      tags: [Documents]
      summary: Get document metadata
      description: Retrieve document information (not file content)
      security:
        - cookieAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Documents]
      summary: Delete document
      description: Permanently delete document and its file
      security:
        - cookieAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Session token stored in HTTP-only cookie
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        char_count:
          type: integer
        created_at:
          type: string
          format: date-time
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        filename:
          type: string
        file_type:
          type: string
          enum: [pdf, txt, docx]
        file_size:
          type: integer
          description: File size in bytes
        uploaded_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type or code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          nullable: true
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing session token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Please log in to access this resource"
    Forbidden:
      description: Forbidden - User lacks permission to access resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "forbidden"
            message: "You do not have permission to access this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "The requested resource was not found"
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "bad_request"
            message: "Invalid input data"
...
    RateLimitExceeded:
      description: Too many requests - Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "rate_limit_exceeded"
            message: "Too many requests. Please try again later."